+++ Src/EonSharp/Wallet.cs
@@ -1,16 +1,29 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Linq;
+using System.Runtime.CompilerServices;
 using System.Runtime.Serialization;
 using System.Text;
 using System.Threading.Tasks;
+using EonSharp.Generators;
 using EonSharp.Keystore;
 
 namespace EonSharp
 {
-	public class Wallet : ISerializable
-	//: Dictionary<string, object>
+	public class Wallet : ISerializable, INotifyPropertyChanged
 	{
+		#region INotifyPropertyChanged
+
+		public event PropertyChangedEventHandler PropertyChanged;
+		void OnPropertyChanged([CallerMemberName]string propertyname = null)
+		{
+			PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyname));
+		}
+
+		#endregion
+
+
 		/// <summary>
 		/// Serialization of Guid
 		/// </summary>
@@ -31,55 +44,77 @@ namespace EonSharp
 		/// </summary>
 		public IKeystore Keystore { get; set; }
 
+		public PublicAccountGenerator AccountDetails
+		{
+			get => m_accountDetails;
+			private set
+			{
+				m_accountDetails = value;
+				OnPropertyChanged();
+			}
+		}
+		PublicAccountGenerator m_accountDetails;
 
 		public Wallet()
 		{
 		}
-
+		public Wallet(string name, string password)
+		{
+			Name = name;
+			var privatekey = SeedGenerator.NewSeed();
+			Keystore = new KeystoreV1(privatekey, password);
+			AccountDetails = new PublicAccountGenerator(privatekey);
+		}
 		public Wallet(string name, byte[] privateKey, string password)
 		{
 			Name = name;
 			Keystore = new KeystoreV1(privateKey, password);
+			AccountDetails = new PublicAccountGenerator(privateKey);
 		}
 
 
+		#region ISerialization
+
 		public Wallet(SerializationInfo info, StreamingContext context)
-		//: base(info, context)
 		{
 			Id = info.GetString("id");
 			Name = info.GetString("name");
 			Version = info.GetInt32("version");
 			Keystore = info.GetValue("keystore", typeof(KeystoreV1)) as IKeystore;
 		}
-
-		//public override void GetObjectData(SerializationInfo info, StreamingContext context)
 		public void GetObjectData(SerializationInfo info, StreamingContext context)
 		{
-			//base.GetObjectData(info, context);
 			info.AddValue("id", Id);
 			info.AddValue("name", Name);
 			info.AddValue("version", Version);
 			info.AddValue("keystore", Keystore);
 		}
 
-		public Generators.PublicAccountGenerator GetPublicAccountGenerator(string password)
+		#endregion
+
+		public void UnlockAccountDetails(string password)
 		{
-			return new Generators.PublicAccountGenerator(Keystore.GetPrivateKey(password));
+			if (AccountDetails == null)
+			{
+				UnlockAccountDetails(Keystore.GetPrivateKey(password));
+			}
 		}
+
 		public byte[] GetPrivateKey(string password)
 		{
-			return Keystore.GetPrivateKey(password);
+			var pv = Keystore.GetPrivateKey(password);
+			UnlockAccountDetails(pv);
+			return pv;
+		}
+		public byte[] GetExpandedKey(string password)
+		{
+			return PublicKeyPairGenerator.ComputeExpandedPrivateKey(GetPrivateKey(password));
 		}
 		public void GetPrivateAndExpandedKeys(string password, out byte[] privatekey, out byte[] expandedprivatekey)
 		{
 			privatekey = GetPrivateKey(password);
-			expandedprivatekey = Generators.PublicKeyPairGenerator.ComputeExpandedPrivateKey(privatekey);
+			expandedprivatekey = PublicKeyPairGenerator.ComputeExpandedPrivateKey(privatekey);
 		}
-
-		//public bool TryGetValue<T>(string key, out T value)
-		//{
-		//	return this.TryGetValue<T>(key, out value);
-		//}
 
 		public string ToJsonString()
 		{
@@ -95,6 +130,12 @@ namespace EonSharp
 			return Newtonsoft.Json.JsonConvert.DeserializeObject<Wallet>(json);
 		}
 
-
+		void UnlockAccountDetails(byte[] privatekey)
+		{
+			if (AccountDetails == null)
+			{
+				AccountDetails = new PublicAccountGenerator(privatekey);
+			}
+		}
 	}
 }